updated readme to include spacy lang packs

+++ utils/tools.py
@@ -1,7 +1,7 @@
 from torch import zeros
 
 
-def epoch_time(start_time: int, end_time: int):
+def epoch_time(start_time: float, end_time: float):
     elapsed_time = end_time - start_time
     elapsed_mins = int(elapsed_time / 60)
     elapsed_secs = int(elapsed_time - (elapsed_mins * 60))
